fix: include project via https crashes when there is no ref
