--- tests/test-cases/include-project-file-ref/.gitlab-ci.yml
@@ -1,3 +1,0 @@
----
-include:
-  - { project: firecow/gitlab-ci-local-includes, file: ["/.gitlab-module.yml", ".gitlab-module.yml"] }
