--- tests/test-cases/include-project-file-ref/mock-gitlab-module.yml
@@ -1,5 +1,0 @@
----
-deploy-job:
-  stage: deploy
-  script:
-    - echo "Deploy something"
